Add gst_flow LLTng tracepoints.
The gst_flow tracepoints trace the flow of data through the gstreamer pipeline.
They can be used to observe latency of the pipeline and its elements.

The following APIs are traced at entry and exit points:

1. Push APIs:
    gst_pad_chain()
    gst_pad_chain_list()
    gst_pad_push()
    gst_pad_push_list()

2. Pull APIs:
    gst_pad_get_range()
    gst_pad_pull_range()

3. Event APIs:
   gst_pad_push_event()
   gst_pad_send_event()

Additionally, the following internal functions are traced:
   gst_pad_push_data()
   gst_pad_chain_data_unchecked()
   gst_pad_get_range_unchecked()

The data stored for each tracepoint:

    -- a unique integer ID of the invoking thread;
    -- the value of the pointer to the pad.

Additionally, at each entry point, the name of the element containing the pad is traced.
For ghost/proxy pads, the name of the element containing the pad's target is traced instead.

At entry points of push and event APIs, and at successful exit points if pull APIs,
the value of the pointer to the data is traced.

At exit points, the call result is traced.

Signed-off-by: Alexander Barabash <alexander_barabash@mentor.com>