Support for gst_flow_event LLTng tracepoints.
The gst_flow_event tracepoints trace the properties of events sent through the gstreamer pipeline.
This replaces gst_flow_data tracepoints for event data.

The initial support for this kind of tracepoints includes calling appropriate primitives
when the data is passed through the pipeline elements, in the following APIs:

   gst_pad_push_event()
   gst_pad_send_event()

The tracepoints are invoked just before the gst_flow tracepoints.

At this stage, only minimal data is stored for each tracepoint:

    -- a unique integer ID of the invoking thread;
    -- the value of the pointer to the event.
    -- the type of event.

Signed-off-by: Alexander Barabash <alexander_barabash@mentor.com>