gst_flow_data LLTng tracepoints.
The gst_flow_data tracepoints trace the properties of data flowing through the gstreamer pipeline.
They can be used to observe throughput of the pipeline and its elements.

The initial support for this kind of tracepoints includes calling appropriate primitives
when the data is passed through the pipeline elements:

1. At entry points of Push APIs:
    gst_pad_chain()
    gst_pad_chain_list()
    gst_pad_push()
    gst_pad_push_list()

2. At successful exit points of Pull APIs:
    gst_pad_get_range()
    gst_pad_pull_range()

The tracepoints are invoked just before the gst_flow tracepoints.

At this stage, only minimal data is stored for each tracepoint:

    -- a unique integer ID of the invoking thread;
    -- the value of the pointer to the data.
    -- the type of data: buffer or buffer list.

For buffer lists, every buffer in the list is also traced.

Signed-off-by: Alexander Barabash <alexander_barabash@mentor.com>